requirements.txt
# Requirements for AgriMithra RAG Service
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.8
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
transformers==4.30.2
torch==2.0.1
python-multipart==0.0.6
